requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
aiohttp==3.12.14
python-dotenv==1.1.1
asyncpg==0.29.0
SQLAlchemy==2.0.29
sqlalchemy[asyncio]
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.5.0
